ref: moved full handshake conformance tests
